workers/requirements.txt
# --- Essenciais ---
requests
python-dotenv
schedule
libsql-client

# --- Dados e Processamento ---
pandas
lxml
html5lib
beautifulsoup4

# --- Scraping ---
playwright